fix: `AuthLinks.tsx`
create component for the new project and new post links
